requirements.txt
asgiref==3.8.1
bcrypt==4.1.2
BTrees==5.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.11
django-cors-headers==4.3.1
djangorestframework==3.15.1
docutils==0.21
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
multipart==0.2.4
packaging==24.0
persistent==5.2
pycparser==2.22
python-gettext==5.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
RestrictedPython==7.1
six==1.16.0
sqlparse==0.4.4
transaction==4.0
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
zc.lockfile==3.0.post1
ZConfig==4.0
zdaemon==5.0
ZEO==6.0.0
ZODB==6.0
zodbpickle==3.2
PyJWT == 2.8.0